+++ src/pipelineSteps/unitTestEngine/chromeHeadless.ts
@@ -14,6 +14,20 @@ export class ChromeHeadless extends PipelineStepBase {
         return super.condition(uniteConfiguration.unitTestEngine, "ChromeHeadless");
     }
 
+    public async initialise(logger: ILogger,
+                            fileSystem: IFileSystem,
+                            uniteConfiguration: UniteConfiguration,
+                            engineVariables: EngineVariables,
+                            mainCondition: boolean): Promise<number> {
+        if (mainCondition) {
+            if (!super.condition(uniteConfiguration.unitTestRunner, "Karma")) {
+                logger.error("You can only use ChromeHeadless unit test engine when the Unit Test runner is Karma");
+                return 1;
+            }
+        }
+        return 0;
+    }
+
     public async configure(logger: ILogger, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables, mainCondition: boolean): Promise<number> {
         engineVariables.toggleDevDependency(["karma-chrome-launcher"], mainCondition && super.condition(uniteConfiguration.unitTestRunner, "Karma"));