+++ src/pipelineSteps/unitTestEngine/phantomJs.ts
@@ -10,8 +10,22 @@ import { EngineVariables } from "../../engine/engineVariables";
 import { PipelineStepBase } from "../../engine/pipelineStepBase";
 
 export class PhantomJs extends PipelineStepBase {
-    public mainCondition(uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables) : boolean | undefined {
+    public mainCondition(uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables): boolean | undefined {
         return super.condition(uniteConfiguration.unitTestEngine, "PhantomJS");
+    }
+
+    public async initialise(logger: ILogger,
+                            fileSystem: IFileSystem,
+                            uniteConfiguration: UniteConfiguration,
+                            engineVariables: EngineVariables,
+                            mainCondition: boolean): Promise<number> {
+        if (mainCondition) {
+            if (!super.condition(uniteConfiguration.unitTestRunner, "Karma")) {
+                logger.error("You can only use PhantomJS unit test engine when the Unit Test runner is Karma");
+                return 1;
+            }
+        }
+        return 0;
     }
 
     public async configure(logger: ILogger, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables, mainCondition: boolean): Promise<number> {